+++ src/main.tsx
@@ -3,6 +3,9 @@ import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App'
 import './index.css'
+import { initVariantLaunch, isIOS, VARIANT_LAUNCH_KEY } from './ar/providers'
+
+if (isIOS() && VARIANT_LAUNCH_KEY) void initVariantLaunch()
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>